+++ dev/App.tsx
@@ -7,22 +7,24 @@ document.body.style.margin = '0';
 document.body.style.padding = '0';
 
 interface Item {
+  name: string;
   value: number;
 }
 
-const getSubtotal = createSelector(
-  ['shop.items'],
-  (items: Item[]) => {
-    return items.reduce<number>((sum, { value }) => {
-      return sum + value;
-    }, 0);
-  },
-  { memoizer: memoize },
+interface State {
+  shop: {
+    items: Item[];
+    taxPercent: number;
+  };
+}
+
+const getSubtotal = createSelector<State>({ memoize })(['shop.items'], (items: Item[]) =>
+  items.reduce<number>((sum, { value }) => sum + value, 0),
 );
-const getTax = createSelector([getSubtotal, 'shop.taxPercent'], (subtotal: number, taxPercent: number) => {
+const getTax = createSelector<State>()([getSubtotal, 'shop.taxPercent'], (subtotal: number, taxPercent: number) => {
   return subtotal * (taxPercent / 100);
 });
-const getTotal = createSelector([getSubtotal, getTax], (subtotal: number, tax: number) => {
+const getTotal = createSelector<State>()([getSubtotal, getTax], (subtotal: number, tax: number) => {
   return subtotal + tax;
 });
 
@@ -40,30 +42,40 @@ console.log('subtotal: ', getSubtotal(state));
 console.log('tax: ', getTax(state));
 console.log('total: ', getTotal(state));
 
-const getFlattedState = createSelector({
-  items: 'shop.items',
-  subtotal: getSubtotal,
-  tax: getTax,
-  total: getTotal,
-});
-console.log('structured state', getFlattedState(state));
-
-const getFoo = createSelector(
-  ['foo'],
-  (foo: string) => {
-    return {
-      bar: foo,
-    };
-  },
+const getFlattedState = createSelector<[State, { test: string }]>()(
   {
-    deepEqual: true,
+    items: 'shop.items',
+    subtotal: getSubtotal,
+    tax: getTax,
+    total: getTotal,
+    extraAuto: { argIndex: 1, path: 'test' },
+    extraManual: (_, { test }) => test,
+  },
+  (items, subtotal, tax, total, auto, manual) => {
+    return [items, subtotal, tax, total, { auto, manual }];
   },
 );
+const flattenedState = getFlattedState(state, { test: 'extra' });
+console.log('structured state', flattenedState);
 
-console.log('using serializer', getFoo({ foo: 'baz' }));
+const getFoo = createSelector<{ foo: string }>()(['foo'], (foo: string) => {
+  return {
+    bar: foo,
+  };
+});
+
+// console.log('using serializer', getFoo({ foo: 'baz' }));
 console.log(getFoo({ foo: 'baz' }));
 
-const getIdentity = createSelector(['foo.bar.baz[0]']);
+interface IdentityState {
+  foo: {
+    bar: {
+      baz: string[];
+    };
+  };
+}
+
+const getIdentity = createSelector<IdentityState>()(['foo.bar.baz[0]']);
 
 console.log(
   getIdentity({
@@ -75,7 +87,7 @@ console.log(
   }),
 );
 
-const getMultipleParams = createSelector(
+const getMultipleParams = createSelector<[typeof first, typeof second, typeof third]>()(
   ['foo.bar', { path: 'baz', argIndex: 1 }, { path: 0, argIndex: 2 }],
   (bar: string, baz: string, quz: string) => {
     return [bar, baz, quz];
@@ -94,7 +106,7 @@ const third = ['blah'];
 
 console.log('mutiple params', getMultipleParams(first, second, third));
 
-const getStructuredParams = createSelector(
+const getStructuredParams = createSelector()(
   ['first.foo.bar', 'second.baz', 'third[0]'],
   (bar: string, baz: string, quz: string) => {
     return [bar, baz, quz];