+++ src/utils.ts
@@ -1,23 +1,32 @@
-import { deepEqual as isDeeplyEqual } from 'fast-equals';
 import { createIdentity } from 'identitate';
-import type { Path, PathItem } from 'pathington';
+import type { Path as PathArray, PathItem } from 'pathington';
 import { parse } from 'pathington';
-import { createSelectorCreator, lruMemoize } from 'reselect';
+import type { CreateSelectorFunction, CreateSelectorOptions } from 'reselect';
+import { createSelectorCreator, weakMapMemoize } from 'reselect';
 import { INVALID_OBJECT_PATH_MESSAGE, INVALID_PATH_MESSAGE } from './constants.js';
-import type { AnyFn, AnyPath, Options, PathObject } from './internalTypes.js';
+import type {
+  AnyFn,
+  ComputeValue,
+  ManualSelectInput,
+  Path,
+  PathObject,
+  PathStructured,
+  SelectStructuredInputs,
+  StructuredValues,
+} from './internalTypes.js';
 
 // eslint-disable-next-line @typescript-eslint/unbound-method
 const hasOwnProperty = Object.prototype.hasOwnProperty;
 
 /**
- * is the path a functions
+ * Whether the path is a functon
  */
-export function isFunctionPath(path: AnyPath): path is AnyFn {
+export function isManualSelector<Params extends unknown[]>(path: Path<Params>): path is ManualSelectInput<Params> {
   return typeof path === 'function';
 }
 
 /**
- * is the path an object
+ * Whether the path is an object.
  */
 export function isObjectPath(path: any): path is PathObject {
   return !!path && typeof path === 'object' && !Array.isArray(path);
@@ -28,19 +37,20 @@ export function isPathItem(path: any): path is PathItem {
 }
 
 /**
- * based on the path passed, create the identity function for it or return the function itself
+ * Create the identity selector function based on the path passed. If the path item is a function then it is used directly,
+ * otherwise the function is created based on its type.
  */
-export function createIdentitySelector(path: AnyPath) {
-  if (isFunctionPath(path)) {
+export function createIdentitySelector<Params extends unknown[]>(path: Path<Params>) {
+  if (isManualSelector(path)) {
     return path;
   }
 
   if (isObjectPath(path)) {
-    if (hasOwnProperty.call(path, 'path') && hasOwnProperty.call(path, 'argIndex')) {
-      const { argIndex, path: objectPath } = path;
+    if (hasOwnProperty.call(path, 'path')) {
+      const { argIndex = 0, path: objectPath } = path;
 
       const selectorIdentity = createIdentity(argIndex);
-      const parsedPath: Path | null = isPathItem(objectPath)
+      const parsedPath: PathItem[] | null = isPathItem(objectPath)
         ? parse(objectPath)
         : Array.isArray(objectPath) && objectPath.every(isPathItem)
           ? parse(objectPath)
@@ -69,7 +79,10 @@ export function createIdentitySelector(path: AnyPath) {
   throw new TypeError(INVALID_PATH_MESSAGE);
 }
 
-function getDeep<State>(path: Path, state: State) {
+/**
+ * Get the value at the deep location expected from the path. If no match is found, return `undefined`.
+ */
+export function getDeep<State>(path: PathArray, state: State) {
   if (state == null) {
     return;
   }
@@ -90,29 +103,37 @@ function getDeep<State>(path: Path, state: State) {
 }
 
 /**
- * get the creator function to use when generating the selector
+ * Get the creator function to use when generating the selector.
  */
 export function getSelectorCreator({
-  deepEqual = false,
-  isEqual = Object.is,
-  memoizer,
-  memoizerParams = [],
-}: Options): AnyFn {
-  const memoizerFn: AnyFn = memoizer ?? lruMemoize;
-  const equals: AnyFn = deepEqual ? isDeeplyEqual : isEqual;
-
-  return createSelectorCreator(memoizerFn, equals, ...memoizerParams);
+  argsMemoize,
+  argsMemoizeOptions,
+  devModeChecks,
+  memoize,
+  memoizeOptions,
+}: CreateSelectorOptions) {
+  return createSelectorCreator({
+    argsMemoize,
+    argsMemoizeOptions,
+    devModeChecks,
+    memoize: memoize ?? weakMapMemoize,
+    memoizeOptions,
+  });
 }
 
 /**
- * get a standard selector based on the paths and getComputedValue provided
+ * Get a standard selector based on the paths and getComputedValue provided.
  */
-export function getStandardSelector(paths: AnyPath[], selectorCreator: AnyFn, getComputedValue: AnyFn): AnyFn {
+export function getStandardSelector<
+  const Params extends unknown[],
+  const Paths extends Array<Path<Params>>,
+  GetComputedValue extends ComputeValue<Params, Paths, any>,
+>(paths: Paths, selectorCreator: CreateSelectorFunction, getComputedValue: GetComputedValue) {
   return selectorCreator(paths.map(createIdentitySelector), getComputedValue);
 }
 
 /**
- * get the structured object based on the computed selector values
+ * Get the structured object based on the computed selector values.
  */
 export function getStructuredObject(properties: string[]): AnyFn {
   return function structuredObject(...values: any[]) {
@@ -123,12 +144,39 @@ export function getStructuredObject(properties: string[]): AnyFn {
     }, {});
   };
 }
-/**
- * get an object of property => selected value pairs bsaed on paths
- */
-export function getStructuredSelector(paths: Record<string, any>, selectorCreator: AnyFn): AnyFn {
-  const destinationKeys = Object.keys(paths);
-  const selectors = destinationKeys.map((key) => createIdentitySelector(paths[key]));
 
-  return selectorCreator(selectors, getStructuredObject(destinationKeys));
+/**
+ * Get an object of property => selected value pairs based on paths.
+ */
+export function getStructuredSelector(
+  paths: PathStructured<any[]>,
+  selectorCreator: CreateSelectorFunction,
+  getComputedValue: AnyFn,
+) {
+  const selectors = Object.keys(paths).map((key) =>
+    createIdentitySelector(
+      // eslint-disable-next-line @typescript-eslint/no-non-null-assertion
+      paths[key]!,
+    ),
+  );
+
+  return selectorCreator(selectors, getComputedValue);
+}
+
+/**
+ * Get an object of property => selected value pairs based on paths.
+ */
+export function getStructuredIdentitySelector<
+  const Params extends unknown[],
+  const Paths extends PathStructured<Params>,
+>(paths: Paths) {
+  const properties = Object.keys(paths);
+
+  return function structuredObject(...values: StructuredValues<Params, Paths>) {
+    return properties.reduce<Record<string, any>>((structuredObject, property, index) => {
+      structuredObject[property] = values[index];
+
+      return structuredObject;
+    }, {}) as SelectStructuredInputs<Params, Paths>;
+  };
 }